Running info.js instruction requires a [prior] npm install
If a user comes straight to the Digital Property app and runs, what is, an OPTIONAL step to see info about all the Sample Applications in the repo, then npm install must be run (at sample-applications level) before running the 'node info.js' step. Put in the word OPTIONAL, because they user can skip the step and go straight to deploying the sample digital property app. Also, the node info steps should be added back to the main README for sample-applications (will do so next)