Fix vehicle manufacture demo (#1)
* Change COMPOSER-VERSION to 0.16.5

Changed from 'latest' to 0.16.5. The latest tag brings down a version that is not compatable with the demo.

* explicitly set FABRIC_VERSION to hlfv1

The downloadFabric.sh script changed the default from hlvf1 to hlvf11.

* Change VEHICLE-LIFECYCLE-VERSION to VEHICLE-MANUFACTURE-VERSION

Updated to work with PR#112

* Delete .project

* added vehicle-manufacture-network.bna to PAYLOAD

The vehicle-manufacture-network.bna was not being included in the tar command, making the bna installation fail.

* NODE-RED-VERSION set to 'latest'

There is no 'next' version, reset it back to 'latest'

* Include instructions to run with hlfv11

* Revert "explicitly set FABRIC_VERSION to hlfv1"

This reverts commit a2190ea6a45b1dc4962a200aed8d722597f61084.

* Revert "explicitly set FABRIC_VERSION to hlfv1"

This reverts commit a2190ea6a45b1dc4962a200aed8d722597f61084.

Signed-off-by: rmiroballi <rmiroballi@gmail.com>

Signed-off-by: Rob Miroballi <rmiroballi@gmail.com>