Vehicle Manufacture Demo not building.
Vehicle Manufacture demo was not building for hlfv1 or hlvf11.
This fixes #126

The COMPOSER-VERSION in build.sh for hlfv1 was changed from 'latest'
to '0.16.5' to fix a compatibility issue.

The build.sh for hlfv11 had placeholder VEHICLE-LIFECYCLE-VERSION
replaced with VEHICLE-MANUFACTURE-VERSION.
The vehicle-manufacture-network.bna was added to the PAYLOAD of the
install.sh.
NODE-RED-VERSION set to 'latest'.

Added instructions to run the demo against hlfv11 to the README

Signed-off-by: Robert Miroballi <rmiroballi@gmail.com>